Add traits module
